Cloud Functions/image_ingest/requirements.txt
google-cloud-storage
requests
datetime